Call markOpen method for storage
Marathon-storage-tool failed to launch for versions of Marathon that
have this safety in place. Unfortunately, 1.4.9 onward and 1.5.2 onward
have this flag, but prior versions do not. In order to reduce the number
of different implementations for bindings, dynamic dispatch is used to
call the method only if it is defined.